requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
brotli==1.1.0
bs4==0.0.2
cachetools==5.3.3
certifi==2025.1.31
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.7
cssselect==1.3.0
cycler==0.12.1
dataclasses-json==0.6.6
deepdiff==7.0.1
deprecated==1.2.14
dill==0.3.8
distro==1.9.0
duckduckgo-search==6.1.5
effdet==0.4.1
emoji==2.12.1
eval-type-backport==0.2.2
exceptiongroup==1.2.1
extra_streamlit_components==0.1.71
faiss-cpu==1.10.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.14.0
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
gitpython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.131.0
googleapis-common-protos==1.63.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-vision==3.7.2
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-oauth==0.15.1
httpx-sse==0.4.0
huggingface-hub==0.23.2
humanfriendly==10.0
icrawler==0.6.8
idna==3.7
intel-openmp==2021.4.0
iopath==0.1.10
jieba3k==0.35.1
jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.50
langchain-google-community==2.0.1
langchain-openai==0.2.3
langchain-text-splitters==0.3.6
langchain-unstructured==0.1.6
langdetect==1.0.9
langsmith==0.1.136
layoutparser==0.3.4
lxml_html_clean==0.4.1
lxml==5.2.2
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.0
openai==1.52.0
opencv-python==4.9.0.80
ordered-set==4.1.0
orjson==3.10.3
outcome==1.3.0.post0 
packaging==23.2
pandas==2.2.2
pdf2image==1.17.0
pdfminer-six==20231228
pdfplumber==0.11.0
pikepdf==8.15.1
pillow==10.3.0
pillow-heif==0.16.0
portalocker==2.8.2
protobuf==4.25.3
proto-plus==1.23.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.6.0
pydeck==0.9.1
pydub==0.25.1
pygments==2.18.0
pyJWT==2.10.1
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
pyreadline3==3.4.1
pyreqwest-impersonate==0.4.7
pysocks==1.7.1 
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytube==15.0.0
pytz==2024.1
pyyaml==6.0.1
rapidfuzz==3.9.2
readability-lxml==0.8.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0 
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
selenium==4.31.0 
sentence-transformers==3.0.0
sgmllib3k==1.0.0 
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0 
soupsieve==2.5
sqlalchemy==2.0.30
streamlit_oauth==0.1.14
streamlit==1.42.0
sympy==1.12.1
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.3
tinysegmenter==0.3 
tldextract==5.1.3
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
transformers==4.41.1
trio==0.29.0 
trio-websocket==0.12.2 
typing-extensions==4.12.2
typing-inspect==0.4.0
tzdata==2024.1
unstructured==0.14.3
unstructured-client==0.31.4
unstructured-inference==0.7.33
unstructured-pytesseract==0.3.12
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.1
webdriver-manager==4.0.2 
websocket-client==1.8.0 
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
youtube-search-python==1.6.6
youtube-transcript-api==0.6.2
yt-dlp==2025.1.26